--- .github/workflows/style.yml
@@ -1,34 +1,0 @@
-# check that the code style is black
-#
-#name: code style
-#
-#on:
-#  push:
-#    branches: [ main ]
-#  pull_request:
-#    branches: [ main ]
-#  workflow_dispatch:
-#
-#jobs:
-#    setup:
-#      runs-on: ubuntu-latest
-#
-#
-#      steps:
-#        - uses: actions/checkout@v2
-#
-#        - name: Set up Python
-#          uses: actions/setup-python@v2
-#          with:
-#              python-version: 3.10
-#
-#        - name: Install poetry
-#          run: |
-#            python -m pip install --upgrade pip
-#            pip install poetry
-#
-#        - name: Install package
-#          run: poetry install
-#
-#        - name: Check code style
-#          run: poetry run black src tests --check
